api: graph: fix signature of get_partition_num()

+++ examples/c/src/cpu_simple_pattern.c
@@ -245,7 +245,7 @@ int main(int argc, char **argv) {
 
     // Get partition from the optimized graph. Each partition will be composed
     // of a single op (fused or unfused op)
-    uint64_t partitions_num;
+    size_t partitions_num;
     DNNL_GRAPH_CHECK(
             dnnl_graph_graph_get_partition_num(graph, &partitions_num));
     if (partitions_num != 2) {
